#12 Fix the namespaces and licensing info
