+++ src/Easify.AspNetCore/Security/AuthOptions.cs
@@ -14,28 +14,11 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-using System;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-
-namespace Easify.Sample.WebAPI.Controllers
+namespace Easify.AspNetCore.Security
 {
-    [Route("api/[controller]")]
-    public class SampleController : Controller
+    public sealed class AuthOptions
     {
-        public SampleController(ILogger<SampleController> logger)
-        {
-            Log = logger ?? throw new ArgumentNullException(nameof(logger));
-        }
-
-        private ILogger<SampleController> Log { get; }
-
-        [HttpGet]
-        [Authorize]
-        public IActionResult Get()
-        {
-            return Ok();
-        }
+        public AuthenticationMode AuthenticationMode { get; set; } = AuthenticationMode.None;
+        public AuthenticationInfo Authentication { get; set; } = new AuthenticationInfo();
     }
 }
