+++ src/Easify.AspNetCore/Security/Impersonations/ImpersonationBearerDefaults.cs
@@ -14,28 +14,10 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-using System;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-
-namespace Easify.Sample.WebAPI.Controllers
+namespace Easify.AspNetCore.Security.Impersonations
 {
-    [Route("api/[controller]")]
-    public class SampleController : Controller
+    public static class ImpersonationBearerDefaults
     {
-        public SampleController(ILogger<SampleController> logger)
-        {
-            Log = logger ?? throw new ArgumentNullException(nameof(logger));
-        }
-
-        private ILogger<SampleController> Log { get; }
-
-        [HttpGet]
-        [Authorize]
-        public IActionResult Get()
-        {
-            return Ok();
-        }
+        public const string AuthenticationScheme = "ImpersonationBearer";
     }
 }
