+++ src/Easify.RestEase.Client/Easify.RestEase.Client.csproj
@@ -13,6 +13,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.Net.Http.Headers" Version="2.0.1" />
     <PackageReference Include="RestEase" Version="1.4.10" />
   </ItemGroup>
