+++ src/Easify.RestEase/IConfigureRestClient.cs
@@ -14,28 +14,10 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-using System;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-
-namespace Easify.Sample.WebAPI.Controllers
+namespace Easify.RestEase
 {
-    [Route("api/[controller]")]
-    public class SampleController : Controller
+    public interface IConfigureRestClient
     {
-        public SampleController(ILogger<SampleController> logger)
-        {
-            Log = logger ?? throw new ArgumentNullException(nameof(logger));
-        }
-
-        private ILogger<SampleController> Log { get; }
-
-        [HttpGet]
-        [Authorize]
-        public IActionResult Get()
-        {
-            return Ok();
-        }
+        IConfigureRestClient ExcludeAuthorizationHeader();
     }
 }
