+++ src/Easify.RestEase/IRestClientBuilder.cs
@@ -14,6 +14,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+using System;
 using Easify.RestEase.Client;
 
 namespace Easify.RestEase
@@ -21,5 +22,6 @@ namespace Easify.RestEase
     public interface IRestClientBuilder
     {
         T Build<T>(string baseUrl) where T : IRestClient;
+        T Build<T>(string baseUrl, Action<IConfigureRestClient> configure) where T : IRestClient;
     }
 }
