+++ src/Easify.RestEase/RestClientOptions.cs
@@ -14,28 +14,16 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-using System;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-
-namespace Easify.Sample.WebAPI.Controllers
+namespace Easify.RestEase
 {
-    [Route("api/[controller]")]
-    public class SampleController : Controller
+    public class RestClientOptions : IConfigureRestClient
     {
-        public SampleController(ILogger<SampleController> logger)
+        public bool IncludeAuthorizationHeader { get; private set; } = true;
+        
+        public IConfigureRestClient ExcludeAuthorizationHeader()
         {
-            Log = logger ?? throw new ArgumentNullException(nameof(logger));
-        }
-
-        private ILogger<SampleController> Log { get; }
-
-        [HttpGet]
-        [Authorize]
-        public IActionResult Get()
-        {
-            return Ok();
+            IncludeAuthorizationHeader = false;
+            return this;
         }
     }
 }
