+++ src/Samples/Easify.Sample.WebAPI/appsettings.json
@@ -21,7 +21,7 @@
     }
   },
   "AuthOptions": {
-    "AuthenticationMode":  "Impersonated",
+    "AuthenticationMode":  "OAuth2",
     "Authentication": {
       "Authority": "https://login.microsoftonline.com/31d4ce72-dfb2-4be8-b876-3278f8641754",
       "Audience": "f7748175-fc0b-4982-8539-070199f3e768"
